Refactor functions into classes
