+++ index.html
@@ -19,6 +19,6 @@
         <button type="submit" class="add-btn">Add</button>
       </form>
     </section>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
