+++ index.js
@@ -1,31 +1,12 @@
+import Del from './DelClass.js';
+import Book from './BookClass.js';
+
 const inputTitle = document.querySelector('#title');
 const inputAuthor = document.querySelector('#author');
 const addBtn = document.querySelector('.add-btn');
 const booksDiv = document.querySelector('.booksDiv');
 
 let books = [];
-
-// Event Listeners
-
-window.addEventListener('DOMContentLoaded', () => {
-  const storageBooks = JSON.parse(localStorage.getItem('books'));
-  if (storageBooks) {
-    books = storageBooks;
-    UI.renderBooks();
-  }
-});
-
-addBtn.addEventListener('click', (e) => {
-  UI.addBook(e);
-  UI.renderBooks();
-});
-
-booksDiv.addEventListener('click', (e) => {
-  if (e.target.classList.contains('remove-btn')) {
-    e.target.parentElement.remove();
-    Del.removeFromStorage(e.target.parentElement);
-  }
-});
 
 // Class: UI -- To handle UI tasks
 class UI {
@@ -34,10 +15,7 @@ class UI {
     if (this.validateForm()) {
       const title = inputTitle.value;
       const author = inputAuthor.value;
-      const book = {
-        title,
-        author,
-      };
+      const book = new Book(title, author);
       books.push(book);
       localStorage.setItem('books', JSON.stringify(books));
       inputAuthor.value = '';
@@ -76,18 +54,24 @@ class UI {
     return true;
   }
 }
+// Event Listeners
 
-// Class: Del -- To delete books from the local storage
-class Del {
-  static removeFromStorage(book) {
-    const bookTitle = book.querySelector('#Ptitle').textContent;
-    const bookAuthor = book.querySelector('#Pauthor').textContent;
-    books = books.filter(
-      (book) => book.title !== bookTitle && book.author !== bookAuthor
-    );
-    localStorage.setItem('books', JSON.stringify(books));
-    if (booksDiv.innerHTML === '') {
-      booksDiv.classList.remove('border');
-    }
+window.addEventListener('DOMContentLoaded', () => {
+  const storageBooks = JSON.parse(localStorage.getItem('books'));
+  if (storageBooks) {
+    books = storageBooks;
+    UI.renderBooks();
   }
-}
+});
+
+addBtn.addEventListener('click', (e) => {
+  UI.addBook(e);
+  UI.renderBooks();
+});
+
+booksDiv.addEventListener('click', (e) => {
+  if (e.target.classList.contains('remove-btn')) {
+    e.target.parentElement.remove();
+    Del.removeFromStorage(e.target.parentElement);
+  }
+});
